feat: divide repository and Service
