--- src/main/java/com/sparta/calender/repository/CalenderReporitory.java
@@ -1,4 +1,0 @@
-package com.sparta.calender.repository;
-
-public class CalenderReporitory {
-}
